% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ISOweek.R
\name{ISOweek}
\alias{ISOweek}
\title{Week of the year according to ISO 8601}
\usage{
ISOweek(date)
}
\arguments{
\item{date}{Vector which can be coerced to class \code{Date}.}
}
\value{
A character vector of year and week in format "\code{\%G-W\%V}".
}
\description{
Returns the year and the week of the year of a given date according to ISO 8601.
It is an alternative to \code{strftime(x, "\%G-W\%V")}.
}
\details{
The conversion specifications \code{\%g}, \code{\%G}, \code{\%V}, \code{\%u}
where not available on Windows and MacOS before \code{R 3.1.0}.
According to ISO 8601, the year of the week can differ from the calendar year (see the examples).
}
\examples{
x <- paste(1999:2011, "-12-31", sep = "")
y <- as.Date(x)
data.frame(date = format(y), week = ISOweek(y))
data.frame(date = x, week = ISOweek(x))
}
\author{
Hatto von Hatzfeld \email{hatto@salesianer.de},
  adopted to \R by Uwe Block \email{u.block.mz@gmail.com}
}
\references{
\url{http://www.salesianer.de/util/kalwoch.html}
}
\seealso{
\code{\link{strftime}} for a description of conversion specifications and references
on ISO 8601.

\code{isoWeekYear} in the \code{surveillance} package for an alternative implementation.
}

