% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISOweekday.R
\name{ISOweekday}
\alias{ISOweekday}
\title{Day of week according to ISO 8601}
\usage{
ISOweekday(date)
}
\arguments{
\item{date}{vector which can be coerced to class \code{Date}}
}
\value{
An integer vector of weekdays (1-7, Monday is 1)
}
\description{
Returns the weekday of a given date according to ISO 8601.
It is an alternative to \code{strftime(x, "\%u")}.
}
\details{
The conversion specifications \code{\%g}, \code{\%G}, \code{\%V}, \code{\%u} 
where not available on Windows and MacOS before \code{R 3.1.0}.
}
\examples{
x <- paste(1999:2011, "-12-31", sep = "")
y <- as.Date(x)
data.frame(date = format(y), weekday = ISOweekday(y))
data.frame(date = x, weekday = ISOweekday(x))
}
\author{
Uwe Block \email{u.block.mz@gmail.com}
}
\seealso{
\code{\link{strftime}}
}

